Create CommandWatcher out of the EmberCli watcher
